+++ src/main/java/com/bdaim/crm/entity/LkCrmBusinessEntity.java
@@ -20,8 +20,8 @@ public class LkCrmBusinessEntity {
     private BigDecimal discountRate;
     private BigDecimal totalPrice;
     private String remark;
-    private int createUserId;
-    private Integer ownerUserId;
+    private Long createUserId;
+    private Long ownerUserId;
     private Timestamp createTime;
     private Timestamp updateTime;
     private String batchId;
@@ -35,7 +35,7 @@ public class LkCrmBusinessEntity {
     //权限（1.只读2.读写）
     private Integer power;
     private String ids;
-    private Integer newOwnerUserId;
+    private Long newOwnerUserId;
     private String memberIds;
     private Integer orderNum;
     private String businessIds;
@@ -163,21 +163,21 @@ public class LkCrmBusinessEntity {
 
     @Basic
     @Column(name = "create_user_id")
-    public int getCreateUserId() {
+    public Long getCreateUserId() {
         return createUserId;
     }
 
-    public void setCreateUserId(int createUserId) {
+    public void setCreateUserId(Long createUserId) {
         this.createUserId = createUserId;
     }
 
     @Basic
     @Column(name = "owner_user_id")
-    public Integer getOwnerUserId() {
+    public Long getOwnerUserId() {
         return ownerUserId;
     }
 
-    public void setOwnerUserId(Integer ownerUserId) {
+    public void setOwnerUserId(Long ownerUserId) {
         this.ownerUserId = ownerUserId;
     }
 
@@ -308,11 +308,11 @@ public class LkCrmBusinessEntity {
         this.ids = ids;
     }
     @Transient
-    public Integer getNewOwnerUserId() {
+    public Long getNewOwnerUserId() {
         return newOwnerUserId;
     }
 
-    public void setNewOwnerUserId(Integer newOwnerUserId) {
+    public void setNewOwnerUserId(Long newOwnerUserId) {
         this.newOwnerUserId = newOwnerUserId;
     }
     @Transient
